* [DXFC-101] Actualize all Records and Events for C# - first try

+++ dxf_api/src/extras/TimeConverter.cs
@@ -17,7 +17,7 @@ namespace com.dxfeed.api.extras
             return offset.AddMilliseconds(utcMillis);
         }
 
-        public static int ToUnixTime(DateTime time)
+        public static long ToUnixTime(DateTime time)
         {
             return (time - offset).Milliseconds;
         }
